session情報からuserIdを取得する処理の汎用化、自身のユーザー情報が削除されないよう条件分岐を追加

+++ app/controller/admin_user_controller.go
@@ -63,6 +63,13 @@ func CreateUsers(c *gin.Context) {
 func DeleteUser(c *gin.Context){
 	// URLからuserIdを取得する
 	userId,_ := strconv.Atoi(c.Param("userId"))
+	// 自分のIdとuserIdが異なることを確認する
+	if userId==GetUserIdBySessionId(c) {
+		fmt.Println("自身のユーザー情報は削除できません")
+		errorMessage := ResponseMessage{Message: "ユーザー削除に失敗しました"}
+		c.JSON(500,errorMessage)
+		return
+	}
 	err:=service.DeleteUser(userId)
 	if err != nil {
 		fmt.Println(err)
